Refactor: Refactor codes; Comment, CommentController, CommentService, CommentRequest
