+++ src/main/java/com/sparta/vicky/comment/dto/CommentRequest.java
@@ -2,13 +2,15 @@ package com.sparta.vicky.comment.dto;
 
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
-import lombok.Data;
+import lombok.Getter;
 
-@Data
+@Getter
 public class CommentRequest {
+
     @NotNull
     private Long boardId;
 
     @NotBlank
     private String content;
+
 }
